+++ src/main/resources/app.properties
@@ -1,4 +1,5 @@
 login = root
 password = !Q2w3e4r
-dbURL = jdbc:mysql://localhost:3306/Test
+testDBURL = jdbc:mysql://localhost:3306/Test
+prodDBURL = jdbc:mysql://localhost:3306/Prod
 SQLClassName = com.mysql.cj.jdbc.Driver
