+++ src/test/java/Services/CityServiceTest.java
@@ -1,0 +1,42 @@
+package Services;
+
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class CityServiceTest {
+
+    @Before
+    public void setUp() throws Exception {
+    }
+
+    @After
+    public void tearDown() throws Exception {
+    }
+
+    @Test
+    public void getInstance() {
+    }
+
+    @Test
+    public void saveFromFile() {
+    }
+
+    @Test
+    public void printAll() {
+    }
+
+    @Test
+    public void printAllOrdered() {
+    }
+
+    @Test
+    public void printGroup() {
+    }
+
+    @Test
+    public void printIndexOfMaxValue() {
+    }
+}
